requirements.txt
fastapi
uvicorn
pydantic
python-jose[cryptography]
passlib[bcrypt]
openai
langchain-core
chromadb
langchain-community
langchain-huggingface
langchain-openai
sentence-transformers
python-multipart
sqlalchemy
aiosqlite
python-dotenv
alembic
httpx
requests
typing-extensions